+++ illuin_config/config_loader.py
@@ -1,5 +1,8 @@
+import logging
 from logging.config import BaseConfigurator, ConvertingDict, ConvertingList, ConvertingTuple
-from typing import Any, Dict, Tuple
+from typing import Any, Dict, Tuple, Optional
+
+from illuin_config.converting_mapping import ConvertingMapping
 
 
 class ConfigLoader(BaseConfigurator):
@@ -11,8 +14,10 @@ class ConfigLoader(BaseConfigurator):
     :ivar config: The parsed configuration.
     """
 
+    logger = logging.getLogger(__name__)
+
     def __init__(self, config_dict: Dict) -> None:
-        self._config: ConvertingDict = None
+        self._config: Optional[ConvertingDict] = None
         BaseConfigurator.__init__(self, config_dict)
 
     @property
@@ -26,8 +31,14 @@ class ConfigLoader(BaseConfigurator):
     def convert(self, value: Any) -> Any:
         value = BaseConfigurator.convert(self, value)
         if isinstance(value, ConvertingDict):
+            value = ConvertingMapping(value)
             if "()" in value:
-                return self.configure_custom(value)
+                cls_path = value["()"]
+                try:
+                    return self.configure_custom(value)
+                except Exception as error:
+                    self.logger.error(f"Could not instantiate {cls_path}: {error!r}")
+                    raise
         elif isinstance(value, ConvertingList):
             value = list(map(self.convert, value))
         elif isinstance(value, ConvertingTuple):
@@ -52,7 +63,7 @@ class ConfigLoader(BaseConfigurator):
         match = self.DOT_PATTERN.match(str_to_parse)  # parsed_object.property
         if match:
             remaining_str: str = str_to_parse[match.end():]
-            if not isinstance(parsed_object, dict):
+            if not isinstance(parsed_object, (dict, ConvertingMapping)):
                 return getattr(parsed_object, match.groups()[0]), remaining_str
             return parsed_object[match.groups()[0]], remaining_str
         match = self.INDEX_PATTERN.match(str_to_parse)  # parsed_object[property]